test : 테스트코드의 중복 수정
