+++ src/main/resources/schema.sql
@@ -31,20 +31,20 @@ create table if not exists LINE_STATION
     updated_at  datetime
 );
 
--- INSERT INTO STATION(ID, NAME)
--- VALUES (1, 'TEST_STATION1');
---
--- INSERT INTO STATION(ID, NAME)
--- VALUES (2, 'TEST_STATION2');
---
--- INSERT INTO STATION(ID, NAME)
--- VALUES (3, 'TEST_STATION3');
---
--- INSERT INTO LINE(ID, NAME, COLOR, START_TIME, END_TIME, INTERVAL_TIME)
--- VALUES (1, 'TEST_LINE1', 'bg-pink-600', '10:00:00', '10:00:00', 1);
---
--- INSERT INTO LINE(ID, NAME, COLOR, START_TIME, END_TIME, INTERVAL_TIME)
--- VALUES (2, 'TEST_LINE2', 'bg-green-600', '10:00:00', '10:00:00', 1);
---
--- INSERT INTO LINE_STATION(LINE, SEQUENCE, STATION, DISTANCE, DURATION)
--- VALUES (1, 1, 1, 0, 0);
+INSERT INTO STATION(ID, NAME)
+VALUES (1, 'TEST_STATION1');
+
+INSERT INTO STATION(ID, NAME)
+VALUES (2, 'TEST_STATION2');
+
+INSERT INTO STATION(ID, NAME)
+VALUES (3, 'TEST_STATION3');
+
+INSERT INTO LINE(ID, NAME, COLOR, START_TIME, END_TIME, INTERVAL_TIME)
+VALUES (1, 'TEST_LINE1', 'bg-pink-600', '10:00:00', '10:00:00', 1);
+
+INSERT INTO LINE(ID, NAME, COLOR, START_TIME, END_TIME, INTERVAL_TIME)
+VALUES (2, 'TEST_LINE2', 'bg-green-600', '10:00:00', '10:00:00', 1);
+
+INSERT INTO LINE_STATION(LINE, SEQUENCE, STATION, DISTANCE, DURATION)
+VALUES (1, 1, 1, 0, 0);
