--- src/test/java/wooteco/subway/admin/SubwayAdminApplicationTests.java
@@ -1,13 +1,0 @@
-package wooteco.subway.admin;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SubwayAdminApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
