+++ src/test/java/wooteco/subway/admin/acceptance/LineAcceptanceTest.java
@@ -3,40 +3,14 @@ package wooteco.subway.admin.acceptance;
 import static org.assertj.core.api.Assertions.*;
 
 import java.time.LocalTime;
-import java.time.format.DateTimeFormatter;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.web.server.LocalServerPort;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import org.springframework.test.context.jdbc.Sql;
 
-import io.restassured.RestAssured;
-import io.restassured.specification.RequestSpecification;
 import wooteco.subway.admin.dto.LineResponse;
 
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@Sql("/truncate.sql")
-public class LineAcceptanceTest {
-
-	@LocalServerPort
-	int port;
-
-	public static RequestSpecification given() {
-		return RestAssured.given().log().all();
-	}
-
-	@BeforeEach
-	void setUp() {
-		RestAssured.port = port;
-	}
-
+public class LineAcceptanceTest extends AcceptanceTest {
 	@DisplayName("지하철 노선을 관리한다")
 	@Test
 	void manageLine() {
@@ -62,7 +36,7 @@ public class LineAcceptanceTest {
 		// when
 		LocalTime startTime = LocalTime.of(8, 0);
 		LocalTime endTime = LocalTime.of(22, 0);
-		updateLine(line.getId(), startTime, endTime);
+		updateLine(line.getId(), line.getName(), startTime, endTime);
 		//then
 		LineResponse updatedLine = getLine(line.getId());
 		assertThat(updatedLine.getStartTime()).isEqualTo(startTime);
@@ -75,68 +49,4 @@ public class LineAcceptanceTest {
 		assertThat(linesAfterDelete.size()).isEqualTo(3);
 	}
 
-	private LineResponse getLine(Long id) {
-		return given().when().
-			get("/api/lines/" + id).
-			then().
-			log().all().
-			extract().as(LineResponse.class);
-	}
-
-	private void createLine(String name) {
-		Map<String, String> params = new HashMap<>();
-		params.put("name", name);
-		params.put("color", "무지개색");
-		params.put("startTime", LocalTime.of(5, 30).format(DateTimeFormatter.ISO_LOCAL_TIME));
-		params.put("endTime", LocalTime.of(23, 30).format(DateTimeFormatter.ISO_LOCAL_TIME));
-		params.put("intervalTime", "10");
-
-		given().
-			body(params).
-			contentType(MediaType.APPLICATION_JSON_VALUE).
-			accept(MediaType.APPLICATION_JSON_VALUE).
-			when().
-			post("/api/lines").
-			then().
-			log().all().
-			statusCode(HttpStatus.CREATED.value());
-	}
-
-	private void updateLine(Long id, LocalTime startTime, LocalTime endTime) {
-		Map<String, String> params = new HashMap<>();
-		params.put("name", "바꿀이름");
-		params.put("color", "초록이");
-		params.put("startTime", startTime.format(DateTimeFormatter.ISO_LOCAL_TIME));
-		params.put("endTime", endTime.format(DateTimeFormatter.ISO_LOCAL_TIME));
-		params.put("intervalTime", "10");
-
-		given().
-			body(params).
-			contentType(MediaType.APPLICATION_JSON_VALUE).
-			accept(MediaType.APPLICATION_JSON_VALUE).
-			when().
-			put("/api/lines/" + id).
-			then().
-			log().all().
-			statusCode(HttpStatus.OK.value());
-	}
-
-	private List<LineResponse> getLines() {
-		return
-			given().
-				when().
-				get("/api/lines").
-				then().
-				log().all().
-				extract().
-				jsonPath().getList(".", LineResponse.class);
-	}
-
-	private void deleteLine(Long id) {
-		given().
-			when().
-			delete("/api/lines/" + id).
-			then().
-			log().all();
-	}
 }
