+++ src/test/java/wooteco/subway/admin/acceptance/StationAcceptanceTest.java
@@ -2,38 +2,14 @@ package wooteco.subway.admin.acceptance;
 
 import static org.assertj.core.api.Assertions.*;
 
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.web.server.LocalServerPort;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import org.springframework.test.context.jdbc.Sql;
 
-import io.restassured.RestAssured;
-import io.restassured.specification.RequestSpecification;
 import wooteco.subway.admin.dto.StationResponse;
 
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@Sql("/truncate.sql")
-public class StationAcceptanceTest {
-	@LocalServerPort
-	int port;
-
-	public static RequestSpecification given() {
-		return RestAssured.given().log().all();
-	}
-
-	@BeforeEach
-	void setUp() {
-		RestAssured.port = port;
-	}
-
+public class StationAcceptanceTest extends AcceptanceTest {
 	@DisplayName("지하철역을 관리한다")
 	@Test
 	void manageStation() {
@@ -49,39 +25,5 @@ public class StationAcceptanceTest {
 
 		List<StationResponse> stationsAfterDelete = getStations();
 		assertThat(stationsAfterDelete.size()).isEqualTo(3);
-	}
-
-	private void createStation(String name) {
-		Map<String, String> params = new HashMap<>();
-		params.put("name", name);
-
-		given().
-			body(params).
-			contentType(MediaType.APPLICATION_JSON_VALUE).
-			accept(MediaType.APPLICATION_JSON_VALUE).
-			when().
-			post("/api/stations").
-			then().
-			log().all().
-			statusCode(HttpStatus.CREATED.value());
-	}
-
-	private List<StationResponse> getStations() {
-
-		return given().
-			when().
-			get("/api/stations").
-			then().
-			log().all().
-			extract().
-			jsonPath().getList(".", StationResponse.class);
-	}
-
-	private void deleteStation(Long id) {
-		given().
-			when().
-			delete("/api/stations/" + id).
-			then().
-			log().all();
 	}
 }
